+++ app/Http/Controllers/AnggotaController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers;
 use App\Models\Anggota;
 use Illuminate\Http\Request;
 use App\Http\Requests\UpdateAnggotaRequest;
+use App\Models\AdminDetail;
+use App\Models\Dojo;
 use App\Models\File;
 use App\Models\User;
 use Exception;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
@@ -20,9 +23,17 @@ class AnggotaController extends Controller
      */
     public function index()
     {
-        return view('anggota.index', [
-            'data' => Anggota::all()
-        ]);
+        $send = [];
+        $send['dojo'] = Dojo::all();
+        if (Gate::allows('isAdmin')) {
+            $admin = AdminDetail::where('id_user', Auth::id())->first();
+            $send['data']  = Anggota::where('id_dojo', $admin->id_dojo)->get();
+            $send['admin'] = $admin; 
+            
+         } else {
+            $send['data'] = Anggota::all();
+         }
+        return view('anggota.index', $send);
     }
 
     /**
@@ -68,7 +79,8 @@ class AnggotaController extends Controller
             'tahun_masuk'    => $request->input('tahun_masuk'),
             'sabut_terakhir' => $request->input('sabut_terakhir'),
             'prestasi'       => $request->input('prestasi'),
-            'bring_by'       => Auth::user()->id
+            'bring_by'       => Auth::user()->id,
+            'id_dojo'        => $request->input('id_dojo'),
         ];
 
         $dataUser['password'] = Hash::make($dataUser['password']);
@@ -217,20 +229,24 @@ class AnggotaController extends Controller
      */
     public function destroy($anggota)
     {
+
+        
+
         $anggota = Anggota::where('id', $anggota)->first();
 
         $foto = File::where('id', $$anggota->foto)->first();
 
+        $user = User::where('id', $anggota->bring_by)->count();
+
+        if(!$user > 0){
+            return redirect()->route('anggota.index')->with('failed', 'Data Gagal Dihapus');
+        }
+
         try {
-
             unlink(public_path('file/' .$foto->name));
-
             User::where('id', $anggota->id_user)->delete();
-
             File::where('id', $anggota->foto)->delete();
-
             Anggota::destroy($anggota->id);
-
             return redirect()->route('anggota.index')->with('success', 'Data Berhasil Dihapus');
         } catch (\Throwable $th) {
             return redirect()->route('anggota.index')->with('failed', 'Data Gagal Dihapus');
